Thanks TKPerson
TKPerson is really the only dev of this project and my mentor(not really). I am sorry for not contributing that much I will get more on top of it. The reason for this message is TKPerson has really done all the coding other than the easy stuff and I want to thank him for doing all of that and helping me learn more coding.